ai-learning/rag-faiss/requirements.txt
faiss-cpu
sentence-transformers
transformers
langchain
pypdf
torch
accelerate
langchain_ollama